% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_data.R
\name{is_data_profile}
\alias{is_data_profile}
\title{Is an object a data_profile?}
\usage{
is_data_profile(x)
}
\arguments{
\item{x}{Any object.}
}
\value{
\code{TRUE} if \code{x} has class \code{data_profile}.
}
\description{
Is an object a data_profile?
}
\examples{
is_data_profile(profile_data(iris))
}
