% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.data_profile}
\alias{print.data_profile}
\title{Print a concise overview of a data profile}
\usage{
\method{print}{data_profile}(x, ...)
}
\arguments{
\item{x}{A \code{data_profile} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a concise overview of a data profile
}
\examples{
print(profile_data(iris))
}
